+++ field_friend/automations/navigation/__init__.py
@@ -1,7 +1,7 @@
 
-from .field_navigation import RowsOnFieldNavigation
 from .follow_crops_navigation import FollowCropsNavigation
 from .navigation import Navigation
+from .row_on_field_navigation import RowsOnFieldNavigation
 from .straight_line_navigation import StraightLineNavigation
 
 __all__ = [
