+++ field_friend/automations/navigation/row_on_field_navigation.py
